fix bug

+++ app/dto/apartment_dto.py
@@ -30,6 +30,6 @@ class RemoveStudentToRoom(BaseModel):
 class ClusterStudent(BaseModel):
     apartment_id: str
     room_type_name: str
-    fuzzy_m: int
+    fuzzy_m: float
     max_loop: int
     epsilon: float
